Move telegraf logger definition to own file
