+++ logger.go
@@ -1,0 +1,23 @@
+package telegraf
+
+var Debug bool
+
+// Logger defines an plugin-related interface for logging.
+type Logger interface {
+	// Errorf logs an error message, patterned after log.Printf.
+	Errorf(format string, args ...interface{})
+	// Error logs an error message, patterned after log.Print.
+	Error(args ...interface{})
+	// Debugf logs a debug message, patterned after log.Printf.
+	Debugf(format string, args ...interface{})
+	// Debug logs a debug message, patterned after log.Print.
+	Debug(args ...interface{})
+	// Warnf logs a warning message, patterned after log.Printf.
+	Warnf(format string, args ...interface{})
+	// Warn logs a warning message, patterned after log.Print.
+	Warn(args ...interface{})
+	// Infof logs an information message, patterned after log.Printf.
+	Infof(format string, args ...interface{})
+	// Info logs an information message, patterned after log.Print.
+	Info(args ...interface{})
+}
